requirements_3.13.txt
# Tested on Python 3.13.1
# =======================
# In the terminal use:
#    pip install -r ./requirements_3.13.txt
#
pip
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-CKEditor==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
python-dotenv